Minor refactor, align the different triton layers
