+++ xformers/triton/layer_norm.py
@@ -12,10 +12,163 @@ from typing import Optional
 import torch
 import torch.nn as nn
 import triton
+from torch.cuda.amp import custom_bwd, custom_fwd
 
-from xformers.triton.k_layer_norm import _LayerNorm
+from xformers.triton.k_layer_norm import (
+    layer_norm_bwd_dwdb,
+    layer_norm_bwd_dx_fused,
+    layer_norm_fw,
+    layer_norm_no_affine_bwd,
+    layer_norm_non_affine_fw,
+)
 
+_triton_layernorm_fp16_enabled = False  # NOTE: PyTorch keeps layernorm as fp32
 _triton_registered_warnings = False
+
+
+class _LayerNorm(torch.autograd.Function):
+    @staticmethod
+    @custom_fwd(cast_inputs=torch.float16 if _triton_layernorm_fp16_enabled else None)
+    def forward(ctx, x, weight, bias, eps):
+        # allocate output
+        y = torch.empty_like(x)
+
+        # reshape input data into 2D tensor
+        x_arg = x.reshape(-1, x.shape[-1])
+        M, N = x_arg.shape
+
+        # allocate mean and std, they'll be used in the backward pass
+        mean = torch.empty((M,), dtype=torch.float32, device="cuda")
+        rstd = torch.empty((M,), dtype=torch.float32, device="cuda")
+
+        # Less than 64KB per feature: enqueue fused kernel
+        MAX_FUSED_SIZE = 65536 // x.element_size()
+        BLOCK_SIZE_N = min(MAX_FUSED_SIZE, triton.next_power_of_2(N))
+        if N > BLOCK_SIZE_N:
+            raise RuntimeError("This layer norm doesn't support feature dim >= 64KB.")
+
+        if not x_arg.is_contiguous() or not y.is_contiguous():
+            global _triton_registered_warnings
+            if not _triton_registered_warnings:
+                logging.warning(
+                    "Non-contiguous input tensor found. Making it contiguous,"
+                    + " but could have perf or trainer implications"
+                )
+
+                _triton_registered_warnings = True
+
+            x_arg = x_arg.contiguous()
+            y = y.contiguous()
+
+        # heuristics for number of warps.
+        num_warps = min(max(BLOCK_SIZE_N // 256, 1), 8)
+
+        # enqueue kernel
+        # fmt: off
+        if weight is None:
+            layer_norm_non_affine_fw[(M,)](
+                x_arg, y, mean, rstd,
+                x_arg.stride(0),
+                N,
+                eps,
+                num_warps=num_warps,
+                BLOCK_SIZE_N=BLOCK_SIZE_N
+            )
+        else:
+            layer_norm_fw[(M,)](
+                x_arg, y, weight, bias, mean, rstd,
+                x_arg.stride(0),
+                N,
+                eps,
+                num_warps=num_warps,
+                BLOCK_SIZE_N=BLOCK_SIZE_N
+            )
+        # fmt: on
+
+        ctx.save_for_backward(y, rstd, weight, bias)
+        ctx.BLOCK_SIZE_N = BLOCK_SIZE_N
+        ctx.num_warps = num_warps
+        ctx.eps = eps
+        ctx.N = N
+
+        return y.reshape_as(x)
+
+    @staticmethod
+    @custom_bwd
+    def backward(
+        ctx, dy
+    ):  # pragma: no cover  # this is covered, but called directly from C++
+        y, var, weight, bias = ctx.saved_tensors
+
+        # heuristics for amount of parallel reduction stream for DG/DB
+        N = y.size(-1)
+        GROUP_SIZE_M = 64
+        if N <= 8192:
+            GROUP_SIZE_M = 96
+        if N <= 4096:
+            GROUP_SIZE_M = 128
+        if N <= 1024:
+            GROUP_SIZE_M = 256
+
+        # flatten the batch dimension, if any.
+        # We're interested in 'samples' x norm_dimension
+        y = y.reshape(-1, y.size(-1))
+        M, N = y.size()
+
+        # allocate output
+        locks = torch.zeros(2 * GROUP_SIZE_M, dtype=torch.int32, device="cuda")
+        t_args = {"dtype": y.dtype, "device": y.device}
+        _dw = torch.empty((GROUP_SIZE_M, y.size(-1)), **t_args)
+        _db = torch.empty((GROUP_SIZE_M, y.size(-1)), **t_args)
+        dw = torch.empty((y.size(-1),), **t_args)
+        db = torch.empty((y.size(-1),), **t_args)
+        dy = dy.contiguous()
+        dx = torch.empty_like(dy)
+
+        # Check the tensor shapes and layouts
+        # we suppose in the kernel that they have the same size and are contiguous
+        assert (
+            dx.numel() == y.numel()
+        ), "Something is wrong in the backward graph, possibly because of an inplace operation after the layernorm"
+
+        # enqueue kernel using forward pass heuristics
+        # also compute partial sums for DW and DB
+
+        # fmt: off
+        meta = {"BLOCK_SIZE_N": ctx.BLOCK_SIZE_N,
+                "GROUP_SIZE_M": GROUP_SIZE_M,
+                "num_warps": ctx.num_warps}
+        if weight is None:
+            layer_norm_no_affine_bwd[(M,)](dx, dy, y, var, y.stride(0), N, **meta)
+            return dx, None, None, None
+
+        layer_norm_bwd_dx_fused[(M,)](
+            dx, dy, _dw, _db,
+            y, weight, bias, var,
+            locks,
+            y.stride(0),
+            N,
+            **meta
+        )
+
+        # fmt: on
+
+        def grid(meta):
+            return [triton.cdiv(N, meta["BLOCK_SIZE_N"])]
+
+        # accumulate partial sums in separate kernel
+        # fmt: off
+        layer_norm_bwd_dwdb[grid](
+            _dw, _db, dw, db,
+            GROUP_SIZE_M,
+            N,
+            BLOCK_SIZE_M=32,
+            BLOCK_SIZE_N=128
+        )
+        # fmt: on
+
+        dx = dx.reshape_as(dy)
+        return dx, dw, db, None
 
 
 class FusedLayerNorm(nn.Module):
